hto fix, changed int to float / function

+++ mean.c
@@ -2,5 +2,5 @@
 #include <stdio.h>
 
 float mean(int a, int b) {
-	return (a + b) / 2;
+	return (a + b) / (float) 2;
 }
